Refactor header file structure and includes
Moved header files to a unified include directory and updated include paths in source files accordingly. Removed obsolete header files and ensured all references are consistent across the project. This improves organization and maintainability of the codebase.
