--- Gary2/include/library.h
@@ -1,18 +1,0 @@
-#ifndef LIBRARY_H
-#define LIBRARY_H
-
-#include <kipr/wombat.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <stdint.h>
-#include <string.h>
-#include <math.h>
-#include <unistd.h>
-#include <time.h>
-
-#include "i2c_comm.h"
-#include "mpu9250.h"
-#include "imu_filter.h"
-#include "motor_control.h"
-
-#endif
